Tratamento de erro Customer Not Found

+++ src/modules/customer/useCases/delete-customer.usecase.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, NotFoundException } from "@nestjs/common";
 import { CustomerRepository } from "../repositories/customer.repository";
 
 
@@ -7,6 +7,11 @@ export class DeleteCustomerUseCase {
     constructor(private readonly customerRepository: CustomerRepository) {}
 
     async deleteCustomer(id: string) {
+        const customerExist = await this.customerRepository.findOne(id);
+        if (!customerExist) {
+            throw new NotFoundException("Customer not found.")
+        }
+        
         return await this.customerRepository.deleteById(id);
     }
 }
